// ui/valset-standardizer-frontend/packages/src/api/generated/valset/schemas/transferTargetUpsertCommand.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Java 外部估值表标准化工程，基于 Spring Boot 和 Quartz
 * OpenAPI spec version: v1.0.0
 */
import type { Object } from "./object";

export interface TransferTargetUpsertCommand {
  /** 连接配置。(object) */
  connectionConfig?: Object;
  /** 是否启用。 */
  enabled?: boolean;
  /** 投递目标编码。 */
  targetCode: string;
  /** 投递目标主键。 */
  targetId?: string;
  /** 目标扩展信息。(object) */
  targetMeta?: Object;
  /** 投递目标名称。 */
  targetName: string;
  /** 投递路径模板。 */
  targetPathTemplate?: string;
  /** 投递目标类型。 */
  targetType: string;
}